#14: Create feature styler (https://github.com/Polymap4/polymap4-p4/issues/issue/14)
added pair field abstract, now used for already existing between form
field and new coord form field
